Don’t steal focus
`autofocus` why used to be able to add several notes in a row by keeping
the focus in the text field even after a screen refresh. However, this
stole the focus at every screen refresh, including when loading Anki.

Since we are not calling `mw.reset()` anymore, keeping the focus now
happens naturally.